# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для системы бронирования столиков в ресторанах",
        "title": "Restaurant Booking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/payments": {
            "get": {
                "description": "Get payment history for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Get user payments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Payment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/payments/halyk": {
            "post": {
                "description": "Create payment via Halyk Bank (returns payment URL)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create Halyk Bank payment",
                "parameters": [
                    {
                        "description": "Payment request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PaymentWithURLResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/payments/kaspi": {
            "post": {
                "description": "Create payment via Kaspi (returns payment URL)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create Kaspi payment",
                "parameters": [
                    {
                        "description": "Payment request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PaymentWithURLResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/payments/wallet": {
            "post": {
                "description": "Create payment using wallet balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create wallet payment",
                "parameters": [
                    {
                        "description": "Payment request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Payment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/payments/webhook/halyk": {
            "post": {
                "description": "Webhook endpoint for Halyk Bank payment notifications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Halyk Bank webhook",
                "parameters": [
                    {
                        "description": "Webhook request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.WebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/payments/webhook/kaspi": {
            "post": {
                "description": "Webhook endpoint for Kaspi payment notifications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Kaspi webhook",
                "parameters": [
                    {
                        "description": "Webhook request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.WebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/payments/{id}/refund": {
            "post": {
                "description": "Refund a completed payment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Refund payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Booking": {
            "type": "object",
            "properties": {
                "booking_date": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "guests_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "restaurant": {
                    "$ref": "#/definitions/domain.Restaurant"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "special_note": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.BookingStatus"
                },
                "table": {
                    "$ref": "#/definitions/domain.Table"
                },
                "table_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/domain.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.BookingStatus": {
            "type": "string",
            "enum": [
                "pending",
                "confirmed",
                "cancelled",
                "completed",
                "no_show"
            ],
            "x-enum-varnames": [
                "BookingStatusPending",
                "BookingStatusConfirmed",
                "BookingStatusCancelled",
                "BookingStatusCompleted",
                "BookingStatusNoShow"
            ]
        },
        "domain.CuisineType": {
            "type": "string",
            "enum": [
                "Italian",
                "Chinese",
                "Mexican",
                "Japanese",
                "Indian",
                "French",
                "Kazakh",
                "Turkish",
                "Thai",
                "American",
                "Korean",
                "Cafe",
                "Bar",
                "Fast Food",
                "Vegetarian",
                "Other"
            ],
            "x-enum-varnames": [
                "CuisineTypeItalian",
                "CuisineTypeChinese",
                "CuisineTypeMexican",
                "CuisineTypeJapanese",
                "CuisineTypeIndian",
                "CuisineTypeFrench",
                "CuisineTypeKazakh",
                "CuisineTypeTurkish",
                "CuisineTypeThai",
                "CuisineTypeAmerican",
                "CuisineTypeKorean",
                "CuisineTypeCafe",
                "CuisineTypeBar",
                "CuisineTypeFastFood",
                "CuisineTypeVegetarian",
                "CuisineTypeOther"
            ]
        },
...
        "domain.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "booking": {
                    "$ref": "#/definitions/domain.Booking"
                },
                "booking_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "external_payment_id": {
                    "type": "string"
                },
                "external_payment_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/domain.PaymentMethod"
                },
                "payment_status": {
                    "$ref": "#/definitions/domain.PaymentStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/domain.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.PaymentMethod": {
            "type": "string",
            "enum": [
                "wallet",
                "halyk",
                "kaspi"
            ],
            "x-enum-varnames": [
                "PaymentMethodWallet",
                "PaymentMethodHalyk",
                "PaymentMethodKaspi"
            ]
        },
        "domain.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "completed",
                "failed",
                "refunded"
            ],
            "x-enum-varnames": [
                "PaymentStatusPending",
                "PaymentStatusCompleted",
                "PaymentStatusFailed",
                "PaymentStatusRefunded"
            ]
        },
        "domain.Restaurant": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "average_price": {
                    "type": "integer"
                },
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Booking"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "cuisine_type": {
                    "$ref": "#/definitions/domain.CuisineType"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.RestaurantImage"
                    }
                },
                "instagram": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "managers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.RestaurantManager"
                    }
                },
                "max_combinable_tables": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/domain.User"
                },
                "owner_id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Review"
                    }
                },
                "reviews_count": {
                    "type": "integer"
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Table"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "working_hours": {
                    "$ref": "#/definitions/domain.WorkingHours"
                }
            }
        },
        "domain.RestaurantImage": {
            "type": "object",
            "properties": {
                "cloudinary_public_id": {
                    "type": "string"
                },
                "cloudinary_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_main": {
                    "type": "boolean"
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
        "domain.RestaurantManager": {
            "type": "object",
            "properties": {
                "assigned_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "restaurant": {
                    "$ref": "#/definitions/domain.Restaurant"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/domain.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.Review": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/domain.Booking"
                },
                "booking_id": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "rating": {
                    "type": "integer"
                },
                "restaurant": {
                    "$ref": "#/definitions/domain.Restaurant"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/domain.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.Table": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Booking"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "location_type": {
                    "$ref": "#/definitions/domain.LocationType"
                },
                "max_capacity": {
                    "type": "integer"
                },
                "min_capacity": {
                    "type": "integer"
                },
                "restaurant": {
                    "$ref": "#/definitions/domain.Restaurant"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "table_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "x_position": {
                    "type": "integer"
                },
                "y_position": {
                    "type": "integer"
                }
            }
        },
...
        "domain.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Booking"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_login_at": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "managed_restaurants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.RestaurantManager"
                    }
                },
                "owned_restaurants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Restaurant"
                    }
                },
                "phone": {
                    "type": "string"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Review"
                    }
                },
                "role": {
                    "$ref": "#/definitions/domain.UserRole"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.UserRole": {
            "type": "string",
            "enum": [
                "customer",
                "owner",
                "manager",
                "admin"
            ],
            "x-enum-varnames": [
                "UserRoleCustomer",
                "UserRoleOwner",
                "UserRoleManager",
                "UserRoleAdmin"
            ]
        },
...
        "domain.WorkingHours": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/domain.DaySchedule"
            }
        },
...
        "handler.CreatePaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "user_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 1
                },
                "booking_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
...
        "handler.PaymentWithURLResponse": {
            "type": "object",
            "properties": {
                "external_payment_url": {
                    "type": "string"
                },
                "payment": {
                    "$ref": "#/definitions/domain.Payment"
                }
            }
        },
...
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation successful"
                }
            }
        },
        "handler.WebhookRequest": {
            "type": "object",
            "required": [
                "external_payment_id",
                "status"
            ],
            "properties": {
                "external_payment_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}